% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\name{create_view_from_definition}
\alias{create_view_from_definition}
\title{Create a View in Snowflake based on an existing table or view}
\usage{
create_view_from_definition(
  con,
  from_database,
  from_schema,
  from_object,
  to_database,
  to_schema,
  to_object,
  role_name = NULL,
  warehouse_name = NULL
)
}
\arguments{
\item{con}{The connection established by the \code{snowflake_con()} function.}

\item{from_database}{Source database name containing the table, dynamic table or view to select columns from.}

\item{from_schema}{Source schema name containing the table, dynamic table or view to select columns from.}

\item{from_object}{Source table, dynamic table or view name to select columns from.}

\item{to_database}{Target database name where the view will be created.}

\item{to_schema}{Target schema name where the view will be created.}

\item{to_object}{Target view name to be created or replaced.}

\item{role_name}{The role to use for the session.}

\item{warehouse_name}{The warehouse to use for the session.}
}
\value{
A data.frame with output metadata on the final SQL execution
}
\description{
This function creates or replaces a view in the specified target database
and schema, using the column definitions from a source table.
}
\examples{
 \dontrun{
# Create Snowflake azure token
tk <- get_az_tk('sf')
con <- snowflake_con(token = tk)

# Ingest Example To Snowflake
create_view_from_definition <- function(
    con,
    from_database = "DEV_PERSISTENT",
    from_schema = "EXAMPLE",
    from_object = "TEST",
    to_database "DEV_PRESENTATION",
    to_schema = "EXAMPLE",
    to_object = "TEST",
    role_name = NULL,
    warehouse_name = NULL
)
)
}
}
