% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\name{ingest_to_snowflake}
\alias{ingest_to_snowflake}
\title{Ingest table into Snowflake}
\usage{
ingest_to_snowflake(data, con, database_name, schema_name, table_name)
}
\arguments{
\item{data}{This is the R object that we are importing into snowflake.
This must be a data.frame, data.table or tibble.}

\item{con}{The connection established by the \code{snowflake_con()} function.}

\item{database_name}{The database where the table will be stored.}

\item{schema_name}{The schema for the table to be stored.}

\item{table_name}{The name of the table for Snowflake}
}
\value{
A data.frame with output metadata on the ingestion process.
}
\description{
Ingest table into Snowflake
}
\examples{
 \dontrun{
# Create Snowflake azure token
tk <- get_az_tk('sf')
con <- snowflake_con(token = tk)

# Ingest Example To Snowflake
ingest_to_snowflake(
  data = data.frame(Alpha = c('a', 'b', 'c'), Number = 1:3),
  con = con,
  database_name = 'EXAMPLE_DATASET',
  schema_name = 'EXAMPLE',
  table_name = 'TEST'
)
}
}
