% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\name{snowflake_con}
\alias{snowflake_con}
\title{Create Connection to Snowflake}
\usage{
snowflake_con(server_name = Sys.getenv("snowflake_server"), token)
}
\arguments{
\item{server_name}{This is the name of Snowflake server you are connecting
to. This should include the name, and the server. For example, this value
could be "mysnowflakeserver.australia-east"}

\item{token}{The token generated with the correct Snowflake permissions.
Use \verb{get_az_tk('sf")} to create this token.}
}
\value{
Establishes a connection to the Snowflake server, and stores this
as an object.
}
\description{
Please ensure that you have the Snowflake ODBC driver installed.
This can be found here:
\itemize{
\item https://docs.snowflake.com/en/user-guide/odbc-download.html.
}

Also ensure that your default role in Snowflake has been set correctly.
If you have not, execute the following command in Snowflake:

ALTER USER "MY_USER_ID@EXAMPLE.COM" SET DEFAULT_ROLE = "MY_DEFAULT_ROLE"
}
\examples{
 \dontrun{
# Create Snowflake azure token
tk <- get_az_tk('sf')
con <- snowflake_con(server_name = 'my-example.server-australiaeast', token = tk)
}
}
